requirements.txt
# Dependencias base - funcional en Windows y Linux (desarrollo y producción con SQLite)
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0

# Producción con PostgreSQL (instalar solo en servidor Linux con PostgreSQL)
# pip install psycopg2-binary
# Producción con Gunicorn (solo Linux)
# pip install gunicorn
